+++ coordinator/gscoordinator/coordinator.py
@@ -20,7 +20,6 @@
 
 import argparse
 import atexit
-import datetime
 import functools
 import json
 import logging
@@ -34,11 +33,7 @@ import string
 import sys
 import threading
 import traceback
-import urllib.parse
-import urllib.request
-import zipfile
 from concurrent import futures
-from io import BytesIO
 
 import grpc
 from packaging import version
@@ -46,68 +41,32 @@ from packaging import version
 from gscoordinator.io_utils import StdStreamWrapper
 
 # capture system stdout
+from gscoordinator.launcher import AbstractLauncher
+from gscoordinator.local_launcher import LocalLauncher
+
 sys.stdout = StdStreamWrapper(sys.stdout)
 sys.stderr = StdStreamWrapper(sys.stderr)
 
-from graphscope.client.utils import GRPCUtils
-from graphscope.framework import utils
-from graphscope.framework.dag_utils import create_graph
-from graphscope.framework.dag_utils import create_loader
-from graphscope.framework.errors import AnalyticalEngineInternalError
-from graphscope.framework.graph_utils import normalize_parameter_edges
-from graphscope.framework.graph_utils import normalize_parameter_vertices
-from graphscope.framework.loader import Loader
 from graphscope.framework.utils import PipeMerger
-from graphscope.framework.utils import find_java
-from graphscope.framework.utils import get_tempdir
-from graphscope.framework.utils import normalize_data_type_str
-from graphscope.proto import attr_value_pb2
+from graphscope.framework.utils import s_to_attr, i_to_attr
 from graphscope.proto import coordinator_service_pb2_grpc
-from graphscope.proto import engine_service_pb2_grpc
 from graphscope.proto import error_codes_pb2
-from graphscope.proto import graph_def_pb2
 from graphscope.proto import message_pb2
-from graphscope.proto import op_def_pb2
 from graphscope.proto import types_pb2
 
-from gscoordinator.cluster import KubernetesClusterLauncher
+from gscoordinator.kubernetes_launcher import KubernetesClusterLauncher
 from gscoordinator.dag_manager import DAGManager
 from gscoordinator.dag_manager import GSEngine
-from gscoordinator.dag_manager import split_op_result
-from gscoordinator.launcher import LocalLauncher
 from gscoordinator.monitor import Monitor
-from gscoordinator.object_manager import GraphMeta
-from gscoordinator.object_manager import GremlinResultSet
 from gscoordinator.object_manager import InteractiveQueryManager
 from gscoordinator.object_manager import LearningInstanceManager
-from gscoordinator.object_manager import LibMeta
 from gscoordinator.object_manager import ObjectManager
-from gscoordinator.utils import ANALYTICAL_ENGINE_JAVA_INIT_CLASS_PATH
-from gscoordinator.utils import ANALYTICAL_ENGINE_JAVA_JVM_OPTS
-from gscoordinator.utils import GRAPHSCOPE_HOME
-from gscoordinator.utils import INTERACTIVE_ENGINE_THREADS_PER_WORKER
-from gscoordinator.utils import RESOURCE_DIR_NAME
-from gscoordinator.utils import WORKSPACE
 from gscoordinator.utils import check_gremlin_server_ready
-from gscoordinator.utils import compile_app
-from gscoordinator.utils import compile_graph_frame
 from gscoordinator.utils import create_single_op_dag
-from gscoordinator.utils import dump_string
-from gscoordinator.utils import get_app_sha256
-from gscoordinator.utils import get_graph_sha256
-from gscoordinator.utils import get_lib_path
-from gscoordinator.utils import op_pre_process
 from gscoordinator.utils import str2bool
-from gscoordinator.utils import to_maxgraph_schema
+from gscoordinator.utils import GS_GRPC_MAX_MESSAGE_LENGTH
 from gscoordinator.version import __version__
-
-# endpoint of prelaunch analytical engine
-GS_DEBUG_ENDPOINT = os.environ.get("GS_DEBUG_ENDPOINT", "")
-
-# 2 GB
-GS_GRPC_MAX_MESSAGE_LENGTH = 2 * 1024 * 1024 * 1024 - 1
-
-logger = logging.getLogger("graphscope")
+from gscoordinator.op_executor import OperationExecutor
 
 
 def catch_unknown_errors(response_on_error=None, using_yield=False):
@@ -127,7 +86,7 @@ def catch_unknown_errors(response_on_error=None, using_yield=False):
             except Exception as exc:
                 error_message = repr(exc)
                 error_traceback = traceback.format_exc()
-                context.set_code(error_codes_pb2.COORDINATOR_INTERNAL_ERROR)
+                context.set_code(grpc.StatusCode.ABORTED)
                 context.set_details(
                     'Error occurs in handler: "%s", with traceback: ' % error_message
                     + error_traceback
@@ -140,470 +99,161 @@ def catch_unknown_errors(response_on_error=None, using_yield=False):
     return catch_exceptions
 
 
+def config_logging(log_level):
+    """Set log level basic on config.
+    Args:
+        log_level (str): Log level of stdout handler
+    """
+    logging.basicConfig(level=logging.CRITICAL)
+
+    if log_level:
+        log_level = log_level.upper()
+
+    logger = logging.getLogger("graphscope")
+    logger.setLevel(log_level)
+
+    stdout_handler = logging.StreamHandler(sys.stdout)
+    stdout_handler.setLevel(log_level)
+    stdout_handler.addFilter(lambda record: record.levelno <= logging.INFO)
+    stderr_handler = logging.StreamHandler(sys.stderr)
+    stderr_handler.setLevel(logging.WARNING)
+
+    formatter = logging.Formatter(
+        "%(asctime)s [%(levelname)s][%(module)s:%(lineno)d]: %(message)s"
+    )
+    stdout_handler.setFormatter(formatter)
+    stderr_handler.setFormatter(formatter)
+
+    logger.addHandler(stdout_handler)
+    logger.addHandler(stderr_handler)
+
+
+logger = logging.getLogger("graphscope")
+
+
 class CoordinatorServiceServicer(
     coordinator_service_pb2_grpc.CoordinatorServiceServicer
 ):
     """Provides methods that implement functionality of master service server.
     Holding:
-        1. process: the grape-engine process.
+        1. launcher: the engine launcher.
         2. session_id: the handle for a particular session to engine
-        3. vineyard_ipc_socket: returned by grape-engine
-        4. vineyard_rpc_socket: returned by grape-engine
-        5. engine_endpoint: the endpoint of grape-engine
-        6. engine_servicer: grpc connection to grape-engine
-
+        3. object_manager: the object manager for the session
+        4. operation_executor: the operation executor for the session
     """
 
-    def __init__(self, launcher, dangling_timeout_seconds, log_level="INFO"):
-        self._launcher = launcher
+    def __init__(
+        self, launcher: AbstractLauncher, dangling_timeout_seconds, log_level="INFO"
+    ):
+        config_logging(log_level)
 
-        self._request = None
         self._object_manager = ObjectManager()
-        self._grpc_utils = GRPCUtils()
-        self._dangling_detecting_timer = None
-        self._config_logging(log_level)
 
         # only one connection is allowed at the same time
-        # generate session id  when a client connection is established
+        # session id will be generated when connection from client is established
         self._session_id = None
+        self._connected = False
 
-        # launch engines
-        if len(GS_DEBUG_ENDPOINT) > 0:
-            logger.info(
-                "Coordinator will connect to engine with endpoint: " + GS_DEBUG_ENDPOINT
-            )
-            self._launcher._analytical_engine_endpoint = GS_DEBUG_ENDPOINT
-        else:
-            if not self._launcher.start():
-                raise RuntimeError("Coordinator Launching failed.")
-
-        self._launcher_type = self._launcher.type()
-        self._instance_id = self._launcher.instance_id
-        # string of a list of hosts, comma separated
-        self._engine_hosts = self._launcher.hosts
-        self._k8s_namespace = ""
-        if self._launcher_type == types_pb2.K8S:
-            self._k8s_namespace = self._launcher.get_namespace()
-
-        # analytical engine
-        self._analytical_engine_stub = self._create_grpc_stub()
-        self._analytical_engine_config = None
-        self._analytical_engine_endpoint = None
-
-        self._builtin_workspace = os.path.join(WORKSPACE, "builtin")
-        # udf app workspace should be bound to a specific session when client connect.
-        self._udf_app_workspace = None
-        # java class path should contains
-        # 1) java runtime path
-        # 2) add resources, the recents added resource will be placed first.
-        self._java_class_path = ANALYTICAL_ENGINE_JAVA_INIT_CLASS_PATH
-        logger.info("Java initial class path set to: {}".format(self._java_class_path))
-        self._jvm_opts = ANALYTICAL_ENGINE_JAVA_JVM_OPTS
+        self._launcher = launcher
 
         # control log fetching
-        self._streaming_logs = True
+        self._streaming_logs = False
         self._pipe_merged = PipeMerger(sys.stdout, sys.stderr)
 
         # dangling check
         self._dangling_timeout_seconds = dangling_timeout_seconds
-        if self._dangling_timeout_seconds >= 0:
-            self._dangling_detecting_timer = threading.Timer(
-                interval=self._dangling_timeout_seconds,
-                function=self._cleanup,
-                args=(
-                    True,
-                    True,
-                ),
-            )
-            self._dangling_detecting_timer.start()
+        self._dangling_detecting_timer = None
+        self._cleanup_instance = False
+        self._set_dangling_timer(cleanup_instance=True)
+
+        self._operation_executor: OperationExecutor = None
 
         # a lock that protects the coordinator
         self._lock = threading.Lock()
-
-        atexit.register(self._cleanup)
+        atexit.register(self.cleanup)
 
     def __del__(self):
-        self._cleanup()
-
-    def _generate_session_id(self):
-        return "session_" + "".join(
-            [random.choice(string.ascii_lowercase) for _ in range(8)]
-        )
-
-    def _config_logging(self, log_level):
-        """Set log level basic on config.
-        Args:
-            log_level (str): Log level of stdout handler
-        """
-        logging.basicConfig(level=logging.CRITICAL)
-
-        if log_level:
-            log_level = log_level.upper()
-
-        logger = logging.getLogger("graphscope")
-        logger.setLevel(log_level)
-
-        stdout_handler = logging.StreamHandler(sys.stdout)
-        stdout_handler.setLevel(log_level)
-        stdout_handler.addFilter(lambda record: record.levelno <= logging.INFO)
-        stderr_handler = logging.StreamHandler(sys.stderr)
-        stderr_handler.setLevel(logging.WARNING)
-
-        formatter = logging.Formatter(
-            "%(asctime)s [%(levelname)s][%(module)s:%(lineno)d]: %(message)s"
-        )
-        stdout_handler.setFormatter(formatter)
-        stderr_handler.setFormatter(formatter)
-
-        logger.addHandler(stdout_handler)
-        logger.addHandler(stderr_handler)
-
-    def ConnectSession(self, request, context):
-        for result in self.ConnectSessionWrapped(request, context):
-            return result
+        self.cleanup()
 
     @Monitor.connectSession
-    def _ConnectSession(self, request, context):
+    def ConnectSession(self, request, context):
         # A session is already connected.
-        if self._request:
+        if self._connected:
             if getattr(request, "reconnect", False):
                 return message_pb2.ConnectSessionResponse(
                     session_id=self._session_id,
                     cluster_type=self._launcher.type(),
                     num_workers=self._launcher.num_workers,
-                    engine_config=json.dumps(self._analytical_engine_config),
-                    pod_name_list=self._engine_hosts.split(","),
-                    namespace=self._k8s_namespace,
+                    host_names=self._launcher.hosts.split(","),
+                    namespace=self._launcher.get_namespace(),
                 )
-            # connect failed, more than one connection at the same time.
-            context.set_code(grpc.StatusCode.ALREADY_EXISTS)
-            context.set_details(
-                "Cannot setup more than one connection at the same time."
-            )
+            else:
+                # connect failed, more than one connection at the same time.
+                context.set_code(grpc.StatusCode.ALREADY_EXISTS)
+                context.set_details(
+                    "Cannot setup more than one connection at the same time."
+                )
+                return message_pb2.ConnectSessionResponse()
+        # check version compatibility from client
+        sv = version.parse(__version__)
+        cv = version.parse(request.version)
+        if sv.major != cv.major or sv.minor != cv.minor:
+            error_msg = f"Version between client and server is inconsistent: {request.version} vs {__version__}"
+            logger.warning(error_msg)
+            context.set_code(grpc.StatusCode.FAILED_PRECONDITION)
+            context.set_details(error_msg)
             return message_pb2.ConnectSessionResponse()
+
         # Connect to serving coordinator.
-        self._key_to_op = {}
-        # dict of op_def_pb2.OpResult
-        self._op_result_pool = {}
-        self._request = request
-        try:
-            self._analytical_engine_config = self._get_engine_config()
-        except grpc.RpcError as e:
-            logger.error(
-                "Get engine config failed, code: %s, details: %s",
-                e.code().name,
-                e.details(),
-            )
-            context.set_code(e.code())
-            context.set_details(e.details())
-            return message_pb2.ConnectSessionResponse()
-        # Generate session id
+        self._connected = True
         self._session_id = self._generate_session_id()
-        self._udf_app_workspace = os.path.join(
-            WORKSPACE, self._instance_id, self._session_id
-        )
-        self._resource_dir = os.path.join(
-            WORKSPACE, self._instance_id, self._session_id, RESOURCE_DIR_NAME
-        )
         self._launcher.set_session_workspace(self._session_id)
+
+        self._operation_executor = OperationExecutor(
+            self._session_id, self._launcher, self._object_manager
+        )
+
+        # Cleanup after timeout seconds
+        self._dangling_timeout_seconds = request.dangling_timeout_seconds
+        # If true, also delete graphscope instance (such as pods) in closing process
+        self._cleanup_instance = request.cleanup_instance
 
         # Session connected, fetch logs via gRPC.
         self._streaming_logs = True
         sys.stdout.drop(False)
-
-        # check version compatibility from client
-        sv = version.parse(__version__)
-        cv = version.parse(self._request.version)
-        if sv.major != cv.major or sv.minor != cv.minor:
-            error_msg = f"Version between client and server is inconsistent: {self._request.version} vs {__version__}"
-            logger.warning(error_msg)
-            context.set_code(error_codes_pb2.CONNECTION_ERROR)
-            context.set_details(error_msg)
-            return message_pb2.ConnectSessionResponse()
-
         return message_pb2.ConnectSessionResponse(
             session_id=self._session_id,
             cluster_type=self._launcher.type(),
             num_workers=self._launcher.num_workers,
-            engine_config=json.dumps(self._analytical_engine_config),
-            pod_name_list=self._engine_hosts.split(","),
-            namespace=self._k8s_namespace,
+            host_names=self._launcher.hosts.split(","),
+            namespace=self._launcher.get_namespace(),
         )
 
-    ConnectSessionWrapped = catch_unknown_errors(message_pb2.ConnectSessionResponse())(
-        _ConnectSession
-    )
+    @Monitor.closeSession
+    def CloseSession(self, request, context):
+        """
+        Disconnect session, note that it won't clean up any resources if self._cleanup_instance is False.
+        """
+        if not self._check_session_consistency(request, context):
+            return message_pb2.CloseSessionResponse()
+
+        self._connected = False
+        self._session_id = None
+
+        self.cleanup(cleanup_instance=self._cleanup_instance, is_dangling=False)
+        self._operation_executor = None
+
+        # Session closed, stop streaming logs
+        sys.stdout.drop(True)
+        self._streaming_logs = False
+        return message_pb2.CloseSessionResponse()
 
     def HeartBeat(self, request, context):
-        for result in self.HeartBeatWrapped(request, context):
-            return result
-
-    def _HeartBeat(self, request, context):
-        if self._request and self._request.dangling_timeout_seconds >= 0:
-            # Reset dangling detect timer
-            if self._dangling_detecting_timer:
-                self._dangling_detecting_timer.cancel()
-
-            self._dangling_detecting_timer = threading.Timer(
-                interval=self._request.dangling_timeout_seconds,
-                function=self._cleanup,
-                args=(
-                    self._request.cleanup_instance,
-                    True,
-                ),
-            )
-            self._dangling_detecting_timer.start()
-
+        self._reset_dangling_timer(self._connected, self._cleanup_instance)
         # analytical engine
-        request = message_pb2.HeartBeatRequest()
-        if self._analytical_engine_stub is None:
-            raise RuntimeError(
-                "Analytical engine is not launched or has already been terminated."
-            )
-        return self._analytical_engine_stub.HeartBeat(request)
-
-    HeartBeatWrapped = catch_unknown_errors(message_pb2.HeartBeatResponse())(_HeartBeat)
-
-    @Monitor.runOnAnalyticalEngine
-    def run_on_analytical_engine(  # noqa: C901
-        self,
-        dag_def: op_def_pb2.DagDef,
-        dag_bodies,
-        loader_op_bodies: dict,
-    ):
-        def _generate_runstep_request(session_id, dag_def, dag_bodies):
-            runstep_requests = []
-            # head
-            runstep_requests.append(
-                message_pb2.RunStepRequest(
-                    head=message_pb2.RunStepRequestHead(
-                        session_id=session_id, dag_def=dag_def
-                    )
-                )
-            )
-            runstep_requests.extend(dag_bodies)
-            for item in runstep_requests:
-                yield item
-
-        # preprocess of op before run on analytical engine
-        for op in dag_def.op:
-            self._key_to_op[op.key] = op
-            op_pre_process(
-                op,
-                self._op_result_pool,
-                self._key_to_op,
-                engine_hosts=self._engine_hosts,
-                engine_config=self._analytical_engine_config,
-                engine_java_class_path=self._java_class_path,  # may be needed in CREATE_GRAPH or RUN_APP
-                engine_jvm_opts=self._jvm_opts,
-            )
-
-            # Handle op that depends on loader (data source)
-            if op.op == types_pb2.CREATE_GRAPH or op.op == types_pb2.ADD_LABELS:
-                for key_of_parent_op in op.parents:
-                    parent_op = self._key_to_op[key_of_parent_op]
-                    if parent_op.op == types_pb2.DATA_SOURCE:
-                        # handle bodies of loader op
-                        if parent_op.key in loader_op_bodies:
-                            dag_bodies.extend(loader_op_bodies[parent_op.key])
-
-            # Compile app or not.
-            if op.op == types_pb2.BIND_APP:
-                op, app_sig, app_lib_path = self._maybe_compile_app(op)
-
-            # Compile graph or not
-            # arrow property graph and project graph need to compile
-            # If engine crashed, we will get a SocketClosed grpc Exception.
-            # In that case, we should notify client the engine is dead.
-            if (
-                (
-                    op.op == types_pb2.CREATE_GRAPH
-                    and op.attr[types_pb2.GRAPH_TYPE].i == graph_def_pb2.ARROW_PROPERTY
-                )
-                or op.op == types_pb2.TRANSFORM_GRAPH
-                or op.op == types_pb2.PROJECT_TO_SIMPLE
-                or op.op == types_pb2.ADD_LABELS
-            ):
-                op = self._maybe_register_graph(op, self._session_id)
-        # generate runstep requests, and run on analytical engine
-        requests = _generate_runstep_request(self._session_id, dag_def, dag_bodies)
-        # response
-        response_head = None
-        response_bodies = []
-        try:
-            responses = self._analytical_engine_stub.RunStep(requests)
-            for response in responses:
-                if response.HasField("head"):
-                    response_head = response
-                else:
-                    response_bodies.append(response)
-        except grpc.RpcError as e:
-            logger.error(
-                "Engine RunStep failed, code: %s, details: %s",
-                e.code().name,
-                e.details(),
-            )
-            if e.code() == grpc.StatusCode.INTERNAL:
-                # TODO: make the stacktrace seperated from normal error messages
-                # Too verbose.
-                if len(e.details()) > 3072:  # 3k bytes
-                    msg = f"{e.details()[:3072]} ... [truncated]"
-                else:
-                    msg = e.details()
-                raise AnalyticalEngineInternalError(msg)
-            else:
-                raise
-
-        # handle result from response stream
-        if response_head is None:
-            raise AnalyticalEngineInternalError(
-                "Missing head from the response stream."
-            )
-        for op_result in response_head.head.results:
-            # record result in coordinator, which doesn't contains large data
-            self._op_result_pool[op_result.key] = op_result
-            # get the op corresponding to the result
-            op = self._key_to_op[op_result.key]
-            # register graph and dump graph schema
-            if op.op in (
-                types_pb2.CREATE_GRAPH,
-                types_pb2.PROJECT_GRAPH,
-                types_pb2.PROJECT_TO_SIMPLE,
-                types_pb2.TRANSFORM_GRAPH,
-                types_pb2.ADD_LABELS,
-                types_pb2.ADD_COLUMN,
-            ):
-                schema_path = os.path.join(
-                    get_tempdir(), op_result.graph_def.key + ".json"
-                )
-                vy_info = graph_def_pb2.VineyardInfoPb()
-                op_result.graph_def.extension.Unpack(vy_info)
-                self._object_manager.put(
-                    op_result.graph_def.key,
-                    GraphMeta(
-                        op_result.graph_def.key,
-                        vy_info.vineyard_id,
-                        op_result.graph_def,
-                        schema_path,
-                    ),
-                )
-                if op_result.graph_def.graph_type == graph_def_pb2.ARROW_PROPERTY:
-                    dump_string(
-                        to_maxgraph_schema(vy_info.property_schema_json),
-                        schema_path,
-                    )
-                    vy_info.schema_path = schema_path
-                    op_result.graph_def.extension.Pack(vy_info)
-            # register app
-            elif op.op == types_pb2.BIND_APP:
-                self._object_manager.put(
-                    app_sig,
-                    LibMeta(op_result.result.decode("utf-8"), "app", app_lib_path),
-                )
-            # unregister graph
-            elif op.op == types_pb2.UNLOAD_GRAPH:
-                self._object_manager.pop(op.attr[types_pb2.GRAPH_NAME].s.decode())
-            # unregister app
-            elif op.op == types_pb2.UNLOAD_APP:
-                self._object_manager.pop(op.attr[types_pb2.APP_NAME].s.decode())
-        return response_head, response_bodies
-
-    @Monitor.runOnInteractiveEngine
-    def run_on_interactive_engine(self, dag_def: op_def_pb2.DagDef):
-        response_head = message_pb2.RunStepResponse(
-            head=message_pb2.RunStepResponseHead()
-        )
-        response_bodies = []
-        for op in dag_def.op:
-            self._key_to_op[op.key] = op
-            op_pre_process(
-                op,
-                self._op_result_pool,
-                self._key_to_op,
-                engine_hosts=self._engine_hosts,
-                engine_config=self._analytical_engine_config,
-            )
-            if op.op == types_pb2.CREATE_INTERACTIVE_QUERY:
-                op_result = self._create_interactive_instance(op)
-            elif op.op == types_pb2.GREMLIN_QUERY:
-                op_result = self._execute_gremlin_query(op)
-            elif op.op == types_pb2.FETCH_GREMLIN_RESULT:
-                op_result = self._fetch_gremlin_result(op)
-            elif op.op == types_pb2.CLOSE_INTERACTIVE_QUERY:
-                op_result = self._close_interactive_instance(op)
-            elif op.op == types_pb2.SUBGRAPH:
-                op_result = self._gremlin_to_subgraph(op)
-            else:
-                raise RuntimeError("Unsupport op type: " + str(op.op))
-            splited_result = split_op_result(op_result)
-            response_head.head.results.append(op_result)
-            for i, chunk in enumerate(splited_result):
-                has_next = True
-                if i + 1 == len(splited_result):
-                    has_next = False
-                response_bodies.append(
-                    message_pb2.RunStepResponse(
-                        body=message_pb2.RunStepResponseBody(
-                            chunk=chunk, has_next=has_next
-                        )
-                    )
-                )
-            # record op result
-            self._op_result_pool[op.key] = op_result
-        return response_head, response_bodies
-
-    def run_on_learning_engine(self, dag_def: op_def_pb2.DagDef):
-        response_head = message_pb2.RunStepResponse(
-            head=message_pb2.RunStepResponseHead()
-        )
-        response_bodies = []
-        for op in dag_def.op:
-            self._key_to_op[op.key] = op
-            op_pre_process(
-                op,
-                self._op_result_pool,
-                self._key_to_op,
-                engine_hosts=self._engine_hosts,
-                engine_config=self._analytical_engine_config,
-            )
-            if op.op == types_pb2.CREATE_LEARNING_INSTANCE:
-                op_result = self._create_learning_instance(op)
-            elif op.op == types_pb2.CLOSE_LEARNING_INSTANCE:
-                op_result = self._close_learning_instance(op)
-            else:
-                raise RuntimeError("Unsupport op type: " + str(op.op))
-            response_head.head.results.append(op_result)
-            self._op_result_pool[op.key] = op_result
-        return response_head, response_bodies
-
-    def run_on_coordinator(
-        self,
-        dag_def: op_def_pb2.DagDef,
-        dag_bodies,
-        loader_op_bodies: dict,
-    ):
-        response_head = message_pb2.RunStepResponse(
-            head=message_pb2.RunStepResponseHead()
-        )
-        response_bodies = []
-        for op in dag_def.op:
-            self._key_to_op[op.key] = op
-            op_pre_process(
-                op,
-                self._op_result_pool,
-                self._key_to_op,
-                engine_hosts=self._engine_hosts,
-                engine_config=self._analytical_engine_config,
-            )
-            if op.op == types_pb2.DATA_SOURCE:
-                op_result = self._process_data_source(op, dag_bodies, loader_op_bodies)
-            elif op.op == types_pb2.DATA_SINK:
-                op_result = self._process_data_sink(op)
-            else:
-                raise RuntimeError("Unsupport op type: " + str(op.op))
-            response_head.head.results.append(op_result)
-            self._op_result_pool[op.key] = op_result
-        return response_head, response_bodies
+        if self._operation_executor is not None:
+            return self._operation_executor.heart_beat(request)
+        return message_pb2.HeartBeatResponse()
 
     def RunStep(self, request_iterator, context):
         with self._lock:
@@ -616,38 +266,37 @@ class CoordinatorServiceServicer(
         loader_op_bodies = {}
 
         # response list for stream
-        responses = []
-        # head
-        responses.append(
+        responses = [
             message_pb2.RunStepResponse(head=message_pb2.RunStepResponseHead())
-        )
+        ]
 
         while not dag_manager.empty():
             run_dag_on, dag, dag_bodies = dag_manager.next_dag()
             error_code = error_codes_pb2.COORDINATOR_INTERNAL_ERROR
-            head = None
-            bodies = None
-
+            head, bodies = None, None
+            # logger.info('dag: %s', dag)
             try:
                 # run on analytical engine
                 if run_dag_on == GSEngine.analytical_engine:
                     # need dag_bodies to load graph from pandas/numpy
                     error_code = error_codes_pb2.ANALYTICAL_ENGINE_INTERNAL_ERROR
-                    head, bodies = self.run_on_analytical_engine(
+                    head, bodies = self._operation_executor.run_on_analytical_engine(
                         dag, dag_bodies, loader_op_bodies
                     )
                 # run on interactive engine
                 elif run_dag_on == GSEngine.interactive_engine:
                     error_code = error_codes_pb2.INTERACTIVE_ENGINE_INTERNAL_ERROR
-                    head, bodies = self.run_on_interactive_engine(dag)
+                    head, bodies = self._operation_executor.run_on_interactive_engine(
+                        dag
+                    )
                 # run on learning engine
                 elif run_dag_on == GSEngine.learning_engine:
                     error_code = error_codes_pb2.LEARNING_ENGINE_INTERNAL_ERROR
-                    head, bodies = self.run_on_learning_engine(dag)
+                    head, bodies = self._operation_executor.run_on_learning_engine(dag)
                 # run on coordinator
                 elif run_dag_on == GSEngine.coordinator:
                     error_code = error_codes_pb2.COORDINATOR_INTERNAL_ERROR
-                    head, bodies = self.run_on_coordinator(
+                    head, bodies = self._operation_executor.run_on_coordinator(
                         dag, dag_bodies, loader_op_bodies
                     )
                 # merge the responses
@@ -663,9 +312,7 @@ class CoordinatorServiceServicer(
                 response_head = responses[0]
                 response_head.head.code = error_code
                 response_head.head.error_msg = (
-                    "Error occurred during preprocessing, The traceback is: {0}".format(
-                        traceback.format_exc()
-                    )
+                    f"Error occurred during RunStep, The traceback is: {traceback.format_exc()}"
                 )
                 response_head.head.full_exception = pickle.dumps(exc)
                 for response in responses:
@@ -678,92 +325,6 @@ class CoordinatorServiceServicer(
         message_pb2.RunStepResponse(head=message_pb2.RunStepResponseHead()), True
     )(_RunStep)
 
-    def _maybe_compile_app(self, op):
-        app_sig = get_app_sha256(op.attr, self._java_class_path)
-        # try to get compiled file from GRAPHSCOPE_HOME/precompiled
-        space = os.path.join(GRAPHSCOPE_HOME, "precompiled", "builtin")
-        app_lib_path = get_lib_path(os.path.join(space, app_sig), app_sig)
-        if not os.path.isfile(app_lib_path):
-            space = self._builtin_workspace
-            if (
-                (types_pb2.GAR in op.attr)
-                or (op.attr[types_pb2.APP_ALGO].s.decode("utf-8").startswith("giraph:"))
-                or op.attr[types_pb2.APP_ALGO].s.decode("utf-8").startswith("java_pie:")
-            ):
-                space = self._udf_app_workspace
-            # try to get compiled file from workspace
-            app_lib_path = get_lib_path(os.path.join(space, app_sig), app_sig)
-            if not os.path.isfile(app_lib_path):
-                # compile and distribute
-                compiled_path = self._compile_lib_and_distribute(
-                    compile_app, app_sig, op
-                )
-                if app_lib_path != compiled_path:
-                    raise RuntimeError(
-                        f"Computed application library path not equal to compiled path, {app_lib_path} versus {compiled_path}"
-                    )
-        op.attr[types_pb2.APP_LIBRARY_PATH].CopyFrom(
-            attr_value_pb2.AttrValue(s=app_lib_path.encode("utf-8"))
-        )
-        return op, app_sig, app_lib_path
-
-    def _maybe_register_graph(self, op, session_id):
-        graph_sig = get_graph_sha256(op.attr)
-        # try to get compiled file from GRAPHSCOPE_HOME/precompiled
-        space = os.path.join(GRAPHSCOPE_HOME, "precompiled", "builtin")
-        graph_lib_path = get_lib_path(os.path.join(space, graph_sig), graph_sig)
-        if not os.path.isfile(graph_lib_path):
-            space = self._builtin_workspace
-            # try to get compiled file from workspace
-            graph_lib_path = get_lib_path(os.path.join(space, graph_sig), graph_sig)
-            if not os.path.isfile(graph_lib_path):
-                # compile and distribute
-                compiled_path = self._compile_lib_and_distribute(
-                    compile_graph_frame, graph_sig, op
-                )
-                if graph_lib_path != compiled_path:
-                    raise RuntimeError(
-                        f"Computed graph library path not equal to compiled path, {graph_lib_path} versus {compiled_path}"
-                    )
-        if graph_sig not in self._object_manager:
-            # register graph
-            op_def = op_def_pb2.OpDef(op=types_pb2.REGISTER_GRAPH_TYPE)
-            op_def.attr[types_pb2.GRAPH_LIBRARY_PATH].CopyFrom(
-                attr_value_pb2.AttrValue(s=graph_lib_path.encode("utf-8"))
-            )
-            op_def.attr[types_pb2.TYPE_SIGNATURE].CopyFrom(
-                attr_value_pb2.AttrValue(s=graph_sig.encode("utf-8"))
-            )
-            op_def.attr[types_pb2.GRAPH_TYPE].CopyFrom(
-                attr_value_pb2.AttrValue(i=op.attr[types_pb2.GRAPH_TYPE].i)
-            )
-            dag_def = op_def_pb2.DagDef()
-            dag_def.op.extend([op_def])
-            try:
-                response_head, _ = self.run_on_analytical_engine(dag_def, [], {})
-            except grpc.RpcError as e:
-                logger.error(
-                    "Register graph failed, code: %s, details: %s",
-                    e.code().name,
-                    e.details(),
-                )
-                if e.code() == grpc.StatusCode.INTERNAL:
-                    raise AnalyticalEngineInternalError(e.details())
-                else:
-                    raise
-            self._object_manager.put(
-                graph_sig,
-                LibMeta(
-                    response_head.head.results[0].result,
-                    "graph_frame",
-                    graph_lib_path,
-                ),
-            )
-        op.attr[types_pb2.TYPE_SIGNATURE].CopyFrom(
-            attr_value_pb2.AttrValue(s=graph_sig.encode("utf-8"))
-        )
-        return op
-
     def FetchLogs(self, request, context):
         while self._streaming_logs:
             try:
@@ -771,7 +332,7 @@ class CoordinatorServiceServicer(
             except queue.Empty:
                 info_message, error_message = "", ""
             except Exception as e:
-                info_message, error_message = "WARNING: failed to read log: %s" % e, ""
+                info_message, error_message = f"WARNING: failed to read log: {e}", ""
 
             if info_message or error_message:
                 if self._streaming_logs:
@@ -780,70 +341,30 @@ class CoordinatorServiceServicer(
                     )
 
     def AddLib(self, request, context):
-        for result in self.AddLibWrapped(request, context):
-            return result
-
-    def _AddLib(self, request, context):
-        if request.session_id != self._session_id:
-            context.set_code(grpc.StatusCode.INVALID_ARGUMENT)
-            context.set_details(
-                f"Session handle not matched, {request.session_id} versus {self._session_id}"
-            )
-        os.makedirs(self._resource_dir, exist_ok=True)
-        gar = request.gar
-        fp = BytesIO(gar)
-        filename = None
-        with zipfile.ZipFile(fp, "r") as zip_ref:
-            zip_ref.extractall(self._resource_dir)
-            logger.info(
-                "Coordinator recieved add lib request contains file {}".format(
-                    zip_ref.namelist()
-                )
-            )
-            if len(zip_ref.namelist()) != 1:
-                raise RuntimeError("Expect only one resource in one gar")
-            filename = zip_ref.namelist()[0]
-        full_filename = os.path.join(self._resource_dir, filename)
-        self._launcher.distribute_file(full_filename)
-        logger.info("Successfully distributed {}".format(full_filename))
-        if full_filename.endswith(".jar"):
-            logger.info("adding lib to java class path since it ends with .jar")
-            self._java_class_path = full_filename + ":" + self._java_class_path
-            logger.info("current java class path: {}".format(self._java_class_path))
+        try:
+            self._operation_executor.add_lib(request)
+        except Exception as e:
+            context.abort(grpc.StatusCode.ABORTED, str(e))
         return message_pb2.AddLibResponse()
 
-    AddLibWrapped = catch_unknown_errors(message_pb2.AddLibResponse())(_AddLib)
-
-    def CloseSession(self, request, context):
-        for result in self.CloseSessionWrapped(request, context):
-            return result
-
-    @Monitor.closeSession
-    def _CloseSession(self, request, context):
-        """
-        Disconnect session, note that it doesn't clean up any resources.
-        """
-        if request.session_id != self._session_id:
-            context.set_code(grpc.StatusCode.INVALID_ARGUMENT)
-            context.set_details(
-                f"Session handle not matched, {request.session_id} versus {self._session_id}"
-            )
-
-        self._cleanup(
-            cleanup_instance=self._request.cleanup_instance, is_dangling=False
+    def CreateAnalyticalInstance(self, request, context):
+        try:
+            self._launcher.start()
+            # create GAE rpc service
+            self._launcher.create_analytical_instance()
+            engine_config = self._operation_executor.get_analytical_engine_config()
+        except grpc.RpcError as e:
+            context.set_code(e.code())
+            context.set_details("Get engine config failed: " + e.details())
+            return message_pb2.CreateAnalyticalInstanceResponse()
+        except Exception as e:
+            context.abort(grpc.StatusCode.ABORTED, str(e))
+            return message_pb2.CreateAnalyticalInstanceResponse()
+        return message_pb2.CreateAnalyticalInstanceResponse(
+            engine_config=json.dumps(engine_config)
         )
-        self._request = None
 
-        # Session closed, stop streaming logs
-        sys.stdout.drop(True)
-        self._streaming_logs = False
-        return message_pb2.CloseSessionResponse()
-
-    CloseSessionWrapped = catch_unknown_errors(message_pb2.CloseSessionResponse())(
-        _CloseSession
-    )
-
-    def _create_interactive_instance(self, op: op_def_pb2.OpDef):
+    def CreateInteractiveInstance(self, request, context):
         def _match_frontend_endpoint(pattern, lines):
             for line in lines.split("\n"):
                 rlt = re.findall(pattern, line)
@@ -851,492 +372,121 @@ class CoordinatorServiceServicer(
                     return rlt[0].strip()
             return ""
 
-        # vineyard object id of graph
-        object_id = op.attr[types_pb2.VINEYARD_ID].i
         # maxgraph endpoint pattern
         FRONTEND_PATTERN = re.compile("(?<=FRONTEND_ENDPOINT:).*$")
+        # maxgraph external endpoint, for clients that are outside of cluster to connect
+        # only available in kubernetes mode, exposed by NodePort or LoadBalancer
         FRONTEND_EXTERNAL_PATTERN = re.compile("(?<=FRONTEND_EXTERNAL_ENDPOINT:).*$")
-        # maxgraph endpoint
-        maxgraph_endpoint = None
-        # maxgraph external endpoint, for client and gremlin function test
-        maxgraph_external_endpoint = None
+
         # create instance
-        proc = self._launcher.create_interactive_instance(op.attr)
+        object_id = request.object_id
+        schema_path = request.schema_path
         try:
+            proc = self._launcher.create_interactive_instance(object_id, schema_path)
+            gie_manager = InteractiveQueryManager(object_id)
+            # Put it to object_manager to ensure it could be killed during coordinator cleanup
+            # If coordinator is shutdown by force when creating interactive instance
+            self._object_manager.put(object_id, gie_manager)
             # 60 seconds is enough, see also GH#1024; try 120
             # already add errs to outs
-            outs, _ = proc.communicate(timeout=120)
+            outs, _ = proc.communicate(timeout=120)  # throws TimeoutError
             return_code = proc.poll()
-            if return_code == 0:
-                # match maxgraph endpoint and check for ready
-                maxgraph_endpoint = _match_frontend_endpoint(FRONTEND_PATTERN, outs)
-                if check_gremlin_server_ready(maxgraph_endpoint):
-                    logger.info(
-                        "build maxgraph frontend %s for graph %ld",
-                        maxgraph_endpoint,
-                        object_id,
-                    )
-                maxgraph_external_endpoint = _match_frontend_endpoint(
-                    FRONTEND_EXTERNAL_PATTERN, outs
+            if return_code != 0:
+                raise RuntimeError(f"Error code: {return_code}, message {outs}")
+            # match maxgraph endpoint and check for ready
+            endpoint = _match_frontend_endpoint(FRONTEND_PATTERN, outs)
+            # coordinator use internal endpoint
+            gie_manager.set_endpoint(endpoint)
+            if check_gremlin_server_ready(endpoint):  # throws TimeoutError
+                logger.info(
+                    "Built interactive frontend %s for graph %ld", endpoint, object_id
                 )
-
-                self._object_manager.put(
-                    op.key,
-                    InteractiveQueryManager(op.key, maxgraph_endpoint, object_id),
-                )
-                endpoint = maxgraph_external_endpoint or maxgraph_endpoint
-                result = {"endpoint": endpoint, "object_id": object_id}
-                return op_def_pb2.OpResult(
-                    code=error_codes_pb2.OK,
-                    key=op.key,
-                    result=json.dumps(result).encode("utf-8"),
-                )
-            raise RuntimeError("Error code: {0}, message {1}".format(return_code, outs))
         except Exception as e:
-            proc.kill()
+            context.set_code(grpc.StatusCode.ABORTED)
+            context.set_details("Create interactive instance failed: " + str(e))
             self._launcher.close_interactive_instance(object_id)
-            raise RuntimeError("Create interactive instance failed.") from e
+            self._object_manager.pop(object_id)
+            return message_pb2.CreateInteractiveInstanceResponse()
+        external_endpoint = _match_frontend_endpoint(FRONTEND_EXTERNAL_PATTERN, outs)
+        # client use external endpoint (k8s mode), or internal endpoint (standalone mode)
+        endpoint = external_endpoint or endpoint
+        return message_pb2.CreateInteractiveInstanceResponse(
+            gremlin_endpoint=endpoint, object_id=object_id
+        )
 
-    def _execute_gremlin_query(self, op: op_def_pb2.OpDef):
-        message = op.attr[types_pb2.GIE_GREMLIN_QUERY_MESSAGE].s.decode()
-        request_options = None
-        if types_pb2.GIE_GREMLIN_REQUEST_OPTIONS in op.attr:
-            request_options = json.loads(
-                op.attr[types_pb2.GIE_GREMLIN_REQUEST_OPTIONS].s.decode()
-            )
-        key_of_parent_op = op.parents[0]
-
-        gremlin_client = self._object_manager.get(key_of_parent_op)
+    def CreateLearningInstance(self, request, context):
+        object_id = request.object_id
+        logger.info("Create learning instance with object id %ld", object_id)
+        handle, config = request.handle, request.config
         try:
-            rlt = gremlin_client.submit(message, request_options=request_options)
+            endpoints = self._launcher.create_learning_instance(
+                object_id, handle, config
+            )
+            self._object_manager.put(object_id, LearningInstanceManager(object_id))
         except Exception as e:
-            raise RuntimeError("Gremlin query failed.") from e
-        self._object_manager.put(op.key, GremlinResultSet(op.key, rlt))
-        return op_def_pb2.OpResult(code=error_codes_pb2.OK, key=op.key)
-
-    def _fetch_gremlin_result(self, op: op_def_pb2.OpDef):
-        fetch_result_type = op.attr[types_pb2.GIE_GREMLIN_FETCH_RESULT_TYPE].s.decode()
-        key_of_parent_op = op.parents[0]
-        result_set = self._object_manager.get(key_of_parent_op).result_set
-        try:
-            if fetch_result_type == "one":
-                rlt = result_set.one()
-            elif fetch_result_type == "all":
-                rlt = result_set.all().result()
-        except Exception as e:
-            raise RuntimeError("Fetch gremlin result failed") from e
-        meta = op_def_pb2.OpResult.Meta(has_large_result=True)
-        return op_def_pb2.OpResult(
-            code=error_codes_pb2.OK,
-            key=op.key,
-            meta=meta,
-            result=pickle.dumps(rlt),
+            context.set_code(grpc.StatusCode.ABORTED)
+            context.set_details("Create learning instance failed: " + str(e))
+            self._launcher.close_learning_instance(object_id)
+            self._object_manager.pop(object_id)
+            return message_pb2.CreateLearningInstanceResponse()
+        return message_pb2.CreateLearningInstanceResponse(
+            object_id=object_id, handle=handle, config=config, endpoints=endpoints
         )
 
-    def _process_data_sink(self, op: op_def_pb2.OpDef):
-        import vineyard
-        import vineyard.io
+    def CloseAnalyticalInstance(self, request, context):
+        context.set_code(grpc.StatusCode.UNIMPLEMENTED)
+        context.set_details("CloseAnalyticalInstance is not implemented")
+        return message_pb2.CloseAnalyticalInstanceResponse()
 
-        storage_options = json.loads(op.attr[types_pb2.STORAGE_OPTIONS].s.decode())
-        fd = op.attr[types_pb2.FD].s.decode()
-        df = op.attr[types_pb2.VINEYARD_ID].s.decode()
-        engine_config = self._analytical_engine_config
-        vineyard_endpoint = engine_config["vineyard_rpc_endpoint"]
-        vineyard_ipc_socket = engine_config["vineyard_socket"]
-        deployment, hosts = self._launcher.get_vineyard_stream_info()
-        dfstream = vineyard.io.open(
-            "vineyard://" + str(df),
-            mode="r",
-            vineyard_ipc_socket=vineyard_ipc_socket,
-            vineyard_endpoint=vineyard_endpoint,
-            deployment=deployment,
-            hosts=hosts,
-        )
-        vineyard.io.open(
-            fd,
-            dfstream,
-            mode="w",
-            vineyard_ipc_socket=vineyard_ipc_socket,
-            vineyard_endpoint=vineyard_endpoint,
-            storage_options=storage_options,
-            deployment=deployment,
-            hosts=hosts,
-        )
-        return op_def_pb2.OpResult(code=error_codes_pb2.OK, key=op.key)
+    def CloseInteractiveInstance(self, request, context):
+        object_id = request.object_id
+        if object_id in self._object_manager:
+            self._object_manager.pop(object_id)
+            try:
+                self._launcher.close_interactive_instance(object_id)
+            except Exception as e:
+                context.set_code(grpc.StatusCode.ABORTED)
+                context.set_details("Close interactive instance failed: " + str(e))
+        return message_pb2.CloseInteractiveInstanceResponse()
 
-    def _process_data_source(
-        self, op: op_def_pb2.OpDef, dag_bodies, loader_op_bodies: dict
-    ):
-        def _spawn_vineyard_io_stream(source, storage_options, read_options):
-            import vineyard
-            import vineyard.io
-
-            engine_config = self._analytical_engine_config
-            vineyard_endpoint = engine_config["vineyard_rpc_endpoint"]
-            vineyard_ipc_socket = engine_config["vineyard_socket"]
-            deployment, hosts = self._launcher.get_vineyard_stream_info()
-            num_workers = self._launcher.num_workers
-            stream_id = repr(
-                vineyard.io.open(
-                    source,
-                    mode="r",
-                    vineyard_endpoint=vineyard_endpoint,
-                    vineyard_ipc_socket=vineyard_ipc_socket,
-                    hosts=hosts,
-                    num_workers=num_workers,
-                    deployment=deployment,
-                    read_options=read_options,
-                    storage_options=storage_options,
-                )
-            )
-            return "vineyard", stream_id
-
-        def _process_loader_func(loader):
-            # loader is type of attr_value_pb2.Chunk
-            protocol = loader.attr[types_pb2.PROTOCOL].s.decode()
-            if protocol in ("hdfs", "hive", "oss", "s3"):
-                source = loader.attr[types_pb2.SOURCE].s.decode()
-                storage_options = json.loads(
-                    loader.attr[types_pb2.STORAGE_OPTIONS].s.decode()
-                )
-                read_options = json.loads(
-                    loader.attr[types_pb2.READ_OPTIONS].s.decode()
-                )
-                new_protocol, new_source = _spawn_vineyard_io_stream(
-                    source, storage_options, read_options
-                )
-                loader.attr[types_pb2.PROTOCOL].CopyFrom(utils.s_to_attr(new_protocol))
-                loader.attr[types_pb2.SOURCE].CopyFrom(utils.s_to_attr(new_source))
-
-        for loader in op.large_attr.chunk_meta_list.items:
-            # handle vertex or edge loader
-            if loader.attr[types_pb2.CHUNK_TYPE].s.decode() == "loader":
-                # set op bodies, this is for loading graph from numpy/pandas
-                op_bodies = []
-                for bodies in dag_bodies:
-                    if bodies.body.op_key == op.key:
-                        op_bodies.append(bodies)
-                loader_op_bodies[op.key] = op_bodies
-                _process_loader_func(loader)
-
-        return op_def_pb2.OpResult(code=error_codes_pb2.OK, key=op.key)
-
-    def _close_interactive_instance(self, op: op_def_pb2.OpDef):
-        try:
-            key_of_parent_op = op.parents[0]
-            gremlin_client = self._object_manager.get(key_of_parent_op)
-            object_id = gremlin_client.object_id
-            proc = self._launcher.close_interactive_instance(object_id)
-            # 60s is enough
-            proc.wait(timeout=60)
-            gremlin_client.close()
-        except Exception as e:
-            raise RuntimeError(
-                f"Failed to close interactive instance {object_id}"
-            ) from e
-        return op_def_pb2.OpResult(
-            code=error_codes_pb2.OK,
-            key=op.key,
-        )
-
-    def _gremlin_to_subgraph(self, op: op_def_pb2.OpDef):
-        gremlin_script = op.attr[types_pb2.GIE_GREMLIN_QUERY_MESSAGE].s.decode()
-        oid_type = op.attr[types_pb2.OID_TYPE].s.decode()
-        request_options = None
-        if types_pb2.GIE_GREMLIN_REQUEST_OPTIONS in op.attr:
-            request_options = json.loads(
-                op.attr[types_pb2.GIE_GREMLIN_REQUEST_OPTIONS].s.decode()
-            )
-        key_of_parent_op = op.parents[0]
-        gremlin_client = self._object_manager.get(key_of_parent_op)
-
-        def create_global_graph_builder(graph_name, num_workers, threads_per_executor):
-            import vineyard
-
-            vineyard_client = vineyard.connect(
-                *self._analytical_engine_config["vineyard_rpc_endpoint"].split(":")
-            )
-
-            instances = [key for key in vineyard_client.meta]
-
-            # duplicate each instances for each thread per worker.
-            chunk_instances = [
-                key for key in instances for _ in range(threads_per_executor)
-            ]
-
-            # build the vineyard::GlobalPGStream
-            metadata = vineyard.ObjectMeta()
-            metadata.set_global(True)
-            metadata["typename"] = "vineyard::htap::GlobalPGStream"
-            metadata["local_stream_chunks"] = threads_per_executor
-            metadata["total_stream_chunks"] = len(chunk_instances)
-
-            # build the parallel stream for edge
-            edge_metadata = vineyard.ObjectMeta()
-            edge_metadata.set_global(True)
-            edge_metadata["typename"] = "vineyard::ParallelStream"
-            edge_metadata["__streams_-size"] = len(chunk_instances)
-
-            # build the parallel stream for vertex
-            vertex_metadata = vineyard.ObjectMeta()
-            vertex_metadata.set_global(True)
-            vertex_metadata["typename"] = "vineyard::ParallelStream"
-            vertex_metadata["__streams_-size"] = len(chunk_instances)
-
-            # NB: we don't respect `num_workers`, instead, we create a substream
-            # on each vineyard instance.
-            #
-            # Such a choice is to handle cases where thet etcd instance still contains
-            # information about dead instances.
-            #
-            # It should be ok, as each engine work will get its own local stream. But,
-            # generally it should be equal to `num_workers`.
-            for worker, instance_id in enumerate(chunk_instances):
-                edge_stream = vineyard.ObjectMeta()
-                edge_stream["typename"] = "vineyard::RecordBatchStream"
-                edge_stream["nbytes"] = 0
-                edge_stream["params_"] = json.dumps(
-                    {
-                        "graph_name": graph_name,
-                        "kind": "edge",
-                    }
-                )
-                edge = vineyard_client.create_metadata(edge_stream, instance_id)
-                vineyard_client.persist(edge.id)
-                edge_metadata.add_member("__streams_-%d" % worker, edge)
-
-                vertex_stream = vineyard.ObjectMeta()
-                vertex_stream["typename"] = "vineyard::RecordBatchStream"
-                vertex_stream["nbytes"] = 0
-                vertex_stream["params_"] = json.dumps(
-                    {
-                        "graph_name": graph_name,
-                        "kind": "vertex",
-                    }
-                )
-                vertex = vineyard_client.create_metadata(vertex_stream, instance_id)
-                vineyard_client.persist(vertex.id)
-                vertex_metadata.add_member("__streams_-%d" % worker, vertex)
-
-                chunk_stream = vineyard.ObjectMeta()
-                chunk_stream["typename"] = "vineyard::htap::PropertyGraphOutStream"
-                chunk_stream["graph_name"] = graph_name
-                chunk_stream["graph_schema"] = "{}"
-                chunk_stream["nbytes"] = 0
-                chunk_stream["stream_index"] = worker
-                chunk_stream.add_member("edge_stream", edge)
-                chunk_stream.add_member("vertex_stream", vertex)
-                chunk = vineyard_client.create_metadata(chunk_stream, instance_id)
-                vineyard_client.persist(chunk.id)
-                metadata.add_member("stream_chunk_%d" % worker, chunk)
-
-            # build the vineyard::GlobalPGStream
-            graph = vineyard_client.create_metadata(metadata)
-            vineyard_client.persist(graph.id)
-            vineyard_client.put_name(graph.id, graph_name)
-
-            # build the parallel stream for edge
-            edge = vineyard_client.create_metadata(edge_metadata)
-            vineyard_client.persist(edge.id)
-            vineyard_client.put_name(edge.id, "__%s_edge_stream" % graph_name)
-
-            # build the parallel stream for vertex
-            vertex = vineyard_client.create_metadata(vertex_metadata)
-            vineyard_client.persist(vertex.id)
-            vineyard_client.put_name(vertex.id, "__%s_vertex_stream" % graph_name)
-
-            return repr(graph.id), repr(edge.id), repr(vertex.id)
-
-        def load_subgraph(
-            graph_name,
-            total_builder_chunks,
-            oid_type,
-            edge_stream_id,
-            vertex_stream_id,
-        ):
-            import vineyard
-
-            # wait all flags been created, see also
-            #
-            # `PropertyGraphOutStream::Initialize(Schema schema)`
-            vineyard_client = vineyard.connect(
-                *self._analytical_engine_config["vineyard_rpc_endpoint"].split(":")
-            )
-
-            # wait for all stream been created by GAIA executor in FFI
-            for worker in range(total_builder_chunks):
-                name = "__%s_%d_streamed" % (graph_name, worker)
-                vineyard_client.get_name(name, wait=True)
-
-            vertices = [Loader(vineyard.ObjectID(vertex_stream_id))]
-            edges = [Loader(vineyard.ObjectID(edge_stream_id))]
-            oid_type = normalize_data_type_str(oid_type)
-            v_labels = normalize_parameter_vertices(vertices, oid_type)
-            e_labels = normalize_parameter_edges(edges, oid_type)
-            loader_op = create_loader(v_labels + e_labels)
-            config = {
-                types_pb2.DIRECTED: utils.b_to_attr(True),
-                types_pb2.OID_TYPE: utils.s_to_attr(oid_type),
-                types_pb2.GENERATE_EID: utils.b_to_attr(False),
-                types_pb2.VID_TYPE: utils.s_to_attr("uint64_t"),
-                types_pb2.IS_FROM_VINEYARD_ID: utils.b_to_attr(False),
-            }
-            new_op = create_graph(
-                self._session_id,
-                graph_def_pb2.ARROW_PROPERTY,
-                inputs=[loader_op],
-                attrs=config,
-            )
-            # spawn a vineyard stream loader on coordinator
-            loader_op_def = loader_op.as_op_def()
-            coordinator_dag = op_def_pb2.DagDef()
-            coordinator_dag.op.extend([loader_op_def])
-            # set the same key from subgraph to new op
-            new_op_def = new_op.as_op_def()
-            new_op_def.key = op.key
-            dag = op_def_pb2.DagDef()
-            dag.op.extend([new_op_def])
-            self.run_on_coordinator(coordinator_dag, [], {})
-            response_head, _ = self.run_on_analytical_engine(dag, [], {})
-            logger.info("subgraph has been loaded")
-            return response_head.head.results[-1]
-
-        # generate a random graph name
-        now_time = datetime.datetime.now().strftime("%Y%m%d%H%M%S")
-        random_num = random.randint(0, 10000000)
-        graph_name = "subgraph-%s-%s" % (str(now_time), str(random_num))
-
-        threads_per_worker = int(
-            os.environ.get("THREADS_PER_WORKER", INTERACTIVE_ENGINE_THREADS_PER_WORKER)
-        )
-
-        if self._launcher_type == types_pb2.HOSTS:
-            # only 1 GIE executor on local cluster
-            executor_workers_num = 1
-            threads_per_executor = self._launcher.num_workers * threads_per_worker
-        else:
-            executor_workers_num = self._launcher.num_workers
-            threads_per_executor = threads_per_worker
-        total_builder_chunks = executor_workers_num * threads_per_executor
-
-        (
-            _graph_builder_id,
-            edge_stream_id,
-            vertex_stream_id,
-        ) = create_global_graph_builder(
-            graph_name, executor_workers_num, threads_per_executor
-        )
-
-        # start a thread to launch the graph
-        pool = futures.ThreadPoolExecutor()
-        subgraph_task = pool.submit(
-            load_subgraph,
-            graph_name,
-            total_builder_chunks,
-            oid_type,
-            edge_stream_id,
-            vertex_stream_id,
-        )
-
-        # add subgraph vertices and edges
-        subgraph_script = "{0}.subgraph('{1}')".format(
-            gremlin_script,
-            graph_name,
-        )
-        gremlin_client.submit(
-            subgraph_script, request_options=request_options
-        ).all().result()
-
-        return subgraph_task.result()
-
-    def _create_learning_instance(self, op: op_def_pb2.OpDef):
-        object_id = op.attr[types_pb2.VINEYARD_ID].i
-        logger.info(
-            "Coordinator create learning instance with object id %ld",
-            object_id,
-        )
-        handle = op.attr[types_pb2.GLE_HANDLE].s.decode("utf-8")
-        config = op.attr[types_pb2.GLE_CONFIG].s.decode("utf-8")
-        endpoints = self._launcher.create_learning_instance(object_id, handle, config)
-        self._object_manager.put(op.key, LearningInstanceManager(op.key, object_id))
-        result = {
-            "handle": handle,
-            "config": config,
-            "endpoints": endpoints,
-            "object_id": object_id,
-        }
-        return op_def_pb2.OpResult(
-            code=error_codes_pb2.OK,
-            key=op.key,
-            result=json.dumps(result).encode("utf-8"),
-        )
-
-    def _close_learning_instance(self, op: op_def_pb2.OpDef):
-        key_of_parent_op = op.parents[0]
-        learning_instance_manager = self._object_manager.get(key_of_parent_op)
-        object_id = learning_instance_manager.object_id
-        logger.info(
-            "Coordinator close learning instance with object id %ld",
-            object_id,
-        )
-        self._launcher.close_learning_instance(object_id)
-        learning_instance_manager.closed = True
-        return op_def_pb2.OpResult(
-            code=error_codes_pb2.OK,
-            key=op.key,
-        )
+    def CloseLearningInstance(self, request, context):
+        object_id = request.object_id
+        if object_id in self._object_manager:
+            self._object_manager.pop(object_id)
+            logger.info("Close learning instance with object id %ld", object_id)
+            try:
+                self._launcher.close_learning_instance(object_id)
+            except Exception as e:
+                context.set_code(grpc.StatusCode.ABORTED)
+                context.set_details("Close learning instance failed: " + str(e))
+        return message_pb2.CloseLearningInstanceResponse()
 
     @Monitor.cleanup
-    def _cleanup(self, cleanup_instance=True, is_dangling=False):
+    def cleanup(self, cleanup_instance=True, is_dangling=False):
         # clean up session resources.
-        for key in self._object_manager.keys():
-            obj = self._object_manager.get(key)
-            obj_type = obj.type
-            unload_type, config = None, None
+        logger.info("Cleaning up resources in coordinator")
+        for _, obj in self._object_manager.items():
+            logger.info("cleaning up %s", obj.type)
+            op_type, config = None, {}
+            if obj.type == "app":
+                op_type = types_pb2.UNLOAD_APP
+                config[types_pb2.APP_NAME] = s_to_attr(obj.key)
+            elif obj.type == "graph":
+                op_type = types_pb2.UNLOAD_GRAPH
+                config[types_pb2.GRAPH_NAME] = s_to_attr(obj.key)
+                # dynamic graph doesn't have a object id
+                if obj.object_id != -1:
+                    config[types_pb2.VINEYARD_ID] = i_to_attr(obj.object_id)
+            elif obj.type == "gie_manager":
+                logger.info("close interactive_instance")
+                self._launcher.close_interactive_instance(obj.object_id)
+            elif obj.type == "gle_manager":
+                self._launcher.close_learning_instance(obj.object_id)
 
-            if obj_type == "app":
-                unload_type = types_pb2.UNLOAD_APP
-                config = {
-                    types_pb2.APP_NAME: attr_value_pb2.AttrValue(
-                        s=obj.key.encode("utf-8")
-                    )
-                }
-            elif obj_type == "graph":
-                unload_type = types_pb2.UNLOAD_GRAPH
-                config = {
-                    types_pb2.GRAPH_NAME: attr_value_pb2.AttrValue(
-                        s=obj.key.encode("utf-8")
-                    )
-                }
-                # dynamic graph doesn't have a vineyard id
-                if obj.vineyard_id != -1:
-                    config[types_pb2.VINEYARD_ID] = attr_value_pb2.AttrValue(
-                        i=obj.vineyard_id
-                    )
-            elif obj_type == "gie_manager":
-                if not obj.closed:
-                    self._close_interactive_instance(
-                        op=op_def_pb2.OpDef(
-                            op=types_pb2.CLOSE_INTERACTIVE_QUERY, parents=[key]
-                        )
-                    )
-
-            elif obj_type == "gle_manager":
-                if not obj.closed:
-                    self._close_learning_instance(
-                        op=op_def_pb2.OpDef(
-                            op=types_pb2.CLOSE_LEARNING_INSTANCE,
-                            parents=[key],
-                        )
-                    )
-
-            if unload_type:
-                dag_def = create_single_op_dag(unload_type, config)
-                request = self._grpc_utils.generate_runstep_requests(
-                    session_id=self._session_id, dag_def=dag_def
-                )
+            if op_type is not None:
+                dag_def = create_single_op_dag(op_type, config)
                 try:
-                    self._analytical_engine_stub.RunStep(request)
+                    self._operation_executor.run_step(dag_def, [])
                 except grpc.RpcError as e:
                     logger.error(
                         "Cleanup failed, code: %s, details: %s",
@@ -1345,82 +495,48 @@ class CoordinatorServiceServicer(
                     )
 
         self._object_manager.clear()
+        self._cancel_dangling_timer()
 
-        self._request = None
+        if cleanup_instance:
+            self._launcher.stop(is_dangling=is_dangling)
 
-        # cancel dangling detect timer
-        if self._dangling_detecting_timer:
+    @staticmethod
+    def _generate_session_id():
+        return "session_" + "".join(
+            [random.choice(string.ascii_lowercase) for _ in range(8)]
+        )
+
+    def _set_dangling_timer(self, cleanup_instance: bool):
+        if self._dangling_timeout_seconds > 0:
+            self._dangling_detecting_timer = threading.Timer(
+                interval=self._dangling_timeout_seconds,
+                function=self.cleanup,
+                args=(
+                    cleanup_instance,
+                    True,
+                ),
+            )
+            self._dangling_detecting_timer.start()
+
+    def _cancel_dangling_timer(self):
+        if self._dangling_detecting_timer is not None:
             self._dangling_detecting_timer.cancel()
             self._dangling_detecting_timer = None
 
-        # close engines
-        if cleanup_instance:
-            self._analytical_engine_stub = None
-            self._analytical_engine_endpoint = None
-            self._launcher.stop(is_dangling=is_dangling)
+    def _reset_dangling_timer(self, reset: bool, cleanup_instance: bool):
+        if reset:
+            self._cancel_dangling_timer()
+            self._set_dangling_timer(cleanup_instance)
 
-        self._session_id = None
-
-    def _create_grpc_stub(self):
-        options = [
-            ("grpc.max_send_message_length", GS_GRPC_MAX_MESSAGE_LENGTH),
-            ("grpc.max_receive_message_length", GS_GRPC_MAX_MESSAGE_LENGTH),
-            ("grpc.max_metadata_size", GS_GRPC_MAX_MESSAGE_LENGTH),
-        ]
-
-        channel = grpc.insecure_channel(
-            self._launcher.analytical_engine_endpoint, options=options
-        )
-        return engine_service_pb2_grpc.EngineServiceStub(channel)
-
-    def _get_engine_config(self):
-        dag_def = create_single_op_dag(types_pb2.GET_ENGINE_CONFIG)
-        try:
-            response_head, _ = self.run_on_analytical_engine(dag_def, [], {})
-        except grpc.RpcError as e:
-            logger.error(
-                "Get engine config failed, code: %s, details: %s",
-                e.code().name,
-                e.details(),
+    def _check_session_consistency(self, request, context):
+        if request.session_id != self._session_id:
+            context.set_code(grpc.StatusCode.INVALID_ARGUMENT)
+            context.set_details(
+                f"Session handle not matched, {request.session_id} versus {self._session_id}"
             )
-            if e.code() == grpc.StatusCode.INTERNAL:
-                raise AnalyticalEngineInternalError(e.details())
-            else:
-                raise
-        config = json.loads(response_head.head.results[0].result.decode("utf-8"))
-        config.update(self._launcher.get_engine_config())
-        # Disable ENABLE_JAVA_SDK when java is not installed on coordinator
-        if config["enable_java_sdk"] == "ON":
-            try:
-                _ = find_java()
-            except RuntimeError:
-                logger.warning(
-                    "Disable java sdk support since java is not installed on coordinator"
-                )
-                config["enable_java_sdk"] = "OFF"
-        return config
-
-    def _compile_lib_and_distribute(self, compile_func, lib_name, op):
-        space = self._builtin_workspace
-        if (
-            (types_pb2.GAR in op.attr)
-            or (op.attr[types_pb2.APP_ALGO].s.decode("utf-8").startswith("giraph:"))
-            or (op.attr[types_pb2.APP_ALGO].s.decode("utf-8").startswith("java_pie:"))
-        ):
-            space = self._udf_app_workspace
-        app_lib_path, java_jar_path, java_ffi_path, app_type = compile_func(
-            space,
-            lib_name,
-            op.attr,
-            self._analytical_engine_config,
-            self._java_class_path,
-        )
-        # for java app compilation, we need to distribute the jar and ffi generated
-        if app_type == "java_pie":
-            self._launcher.distribute_file(java_jar_path)
-            self._launcher.distribute_file(java_ffi_path)
-        self._launcher.distribute_file(app_lib_path)
-        return app_lib_path
+            return False
+        else:
+            return True
 
 
 def parse_sys_args():
@@ -1523,7 +639,7 @@ def parse_sys_args():
         "--k8s_image_pull_secrets",
         type=str,
         default="graphscope",
-        help="A list of comma sparated secrets to pull image.",
+        help="A list of comma separated secrets to pull image.",
     )
     parser.add_argument(
         "--k8s_vineyard_daemonset",
@@ -1535,7 +651,7 @@ def parse_sys_args():
         "--k8s_vineyard_cpu",
         type=float,
         default=1.0,
-        help="CPU cores of vinayard container.",
+        help="CPU cores of vineyard container.",
     )
     parser.add_argument(
         "--k8s_vineyard_mem",
@@ -1645,7 +761,7 @@ def parse_sys_args():
         "--k8s_volumes",
         type=str,
         default="{}",
-        help="A json string spcifies the kubernetes volumes to mount.",
+        help="A json string specifies the kubernetes volumes to mount.",
     )
     parser.add_argument(
         "--timeout_seconds",
@@ -1802,7 +918,7 @@ def launch_graphscope():
 
     # handle SIGTERM signal
     def terminate(signum, frame):
-        coordinator_service_servicer._cleanup()
+        coordinator_service_servicer.cleanup()
 
     signal.signal(signal.SIGTERM, terminate)
 
@@ -1810,7 +926,7 @@ def launch_graphscope():
         # Grpc has handled SIGINT
         server.wait_for_termination()
     except KeyboardInterrupt:
-        coordinator_service_servicer._cleanup()
+        coordinator_service_servicer.cleanup()
 
 
 if __name__ == "__main__":
