+++ coordinator/gscoordinator/learning.py
@@ -19,7 +19,6 @@
 import base64
 import json
 import logging
-import os
 import sys
 
 import graphscope.learning.graphlearn as gl
