+++ coordinator/requirements.txt
@@ -3,9 +3,13 @@ etcd-distro>=3.5.1
 graphscope-client>=0.11.0
 grpcio<=1.43.0,>=1.40.0
 grpcio-tools<=1.43.0,>=1.40.0
-kubernetes
+kubernetes~=12.0.1
 protobuf>=3.15.0,<=3.18.1
-PyYAML
+PyYAML~=5.4.1
 vineyard>=0.5.2; sys_platform != "win32"
 vineyard-io>=0.5.2; sys_platform != "win32"
 prometheus-client>=0.14.1
+
+setuptools~=51.1.2
+packaging~=20.8
+tqdm~=4.62.3
