+++ interactive_engine/executor/store/global_query/Cargo.toml
@@ -11,7 +11,7 @@ itertools = "0.7.8"
 byteorder = "1.3.1"
 serde = "1.0.72"
 libc = "0.2"
-maxgraph-store = { path = "../groot", optional = true }
+maxgraph-store = { path = "../groot" }
 ir_common = {path = "../../ir/common"}
 dyn_type = { path = "../../common/dyn_type" }
