+++ python/graphscope/framework/dag.py
@@ -84,7 +84,7 @@ class Dag(object):
         # assert op is not present in current dag
         for op in ops:
             assert op.key in self._ops_by_key, "%s is not in the dag" % op.key
-            assert not self._ops_by_key[op.key].evaluated, "%is is evaluated" % op.key
+            assert not self._ops_by_key[op.key].evaluated, "%s is evaluated" % op.key
             op_keys.append(op.key)
         op_keys_to_keep = self._bfs_for_reachable_ops(op_keys)
         op_keys_to_keep = sorted(op_keys_to_keep, key=lambda n: self._ops_seq_by_key[n])
