+++ python/graphscope/nx/conftest.py
@@ -26,8 +26,6 @@ import graphscope
 @pytest.fixture(scope="module")
 def graphscope_session():
     graphscope.set_option(show_log=True)
-    graphscope.set_option(initializing_interactive_engine=False)
-
     if os.environ.get("DEPLOYMENT", None) == "standalone":
         sess = graphscope.session(cluster_type="hosts", num_workers=1)
     else:
